Updated TagServiceImpl to be Transacational
Removed final modifiers from all Transnational methods, i.e. add and get. 
Due to Checkstyle failures with the removal of finals, we now have to suppress warnings for DesignForExtension.
We should start using findById instead of getOne due to the fact that getOne simply returns the reference to an entity, which means the entity fields can be null - bad for integration test check.